refactor method.jsx
